feat: [split-world-state-generated-adapters-001] - [Extract initial-structure generated adapters into focused topology helpers]
